Remove cs_high capability
This functionality was removed in the 5.4 kernel line, and this causes
SPIDEV to fail, but the error manifests in luma.core when setting the
cs_high line.

The specialist class cs_gpio_spi still supports setting the CS active
high if necessary.

This issue has already been reported in several places, so this PR
should address these:
* Fixes https://github.com/rm-hull/luma.led_matrix/issues/225
* Fixes https://github.com/rm-hull/luma.oled/issues/292
* Fixes https://github.com/rm-hull/luma.examples/issues/124