tests: do not match exact OS errno strings (#238)
The actual string representation of an errno (like ENOENT) is an OS
implementation detail, and there is no guarantee on it. Also, we are
already checking that the exception is FileNotFoundError, which implies
ENOENT already.

Hence, simplify the assertion checking: check that the actual filename
that triggered the exception was the wrong one that was passed to
load_sprite_table().